refactor: move QueryablesDict class
