+++ eodag/resources/product_types.yml
@@ -2396,7 +2396,7 @@ CAMS_EU_AIR_QUALITY_FORECAST:
   sensorType: ATMOSPHERIC
   license: proprietary
   title: CAMS European air quality forecasts
-  missionStartDate: "2020-08-26T00:00:00Z"
+  missionStartDate: "2021-11-17T00:00:00Z"
 
 CAMS_GFE_GFAS:
   abstract: |
